+++ source/plugins/go/src/utils.go
@@ -63,7 +63,13 @@ func ReadConfiguration(filename string) (map[string]string, error) {
 
 // CreateHTTPClient used to create the client for sending post requests to OMSEndpoint
 func CreateHTTPClient() {
-	cert, err := tls.LoadX509KeyPair(PluginConfiguration["cert_file_path"], PluginConfiguration["key_file_path"])
+	certFilePath := PluginConfiguration["cert_file_path"]
+	keyFilePath := PluginConfiguration["key_file_path"]
+	if IsWindows == false {
+		certFilePath = fmt.Sprintf(certFilePath, WorkspaceID)
+		keyFilePath = fmt.Sprintf(keyFilePath, WorkspaceID)
+	}
+	cert, err := tls.LoadX509KeyPair(certFilePath, keyFilePath)
 	if err != nil {
 		message := fmt.Sprintf("Error when loading cert %s", err.Error())
 		SendException(message)
@@ -111,27 +117,58 @@ func ToString(s interface{}) string {
 }
 
 //mdsdSocketClient to write msgp messages
-func CreateMDSDClient(containerType string) {
-	if MdsdMsgpUnixSocketClient != nil {
-		MdsdMsgpUnixSocketClient.Close()
-		MdsdMsgpUnixSocketClient = nil
-	}
-	/*conn, err := fluent.New(fluent.Config{FluentNetwork:"unix",
-	  FluentSocketPath:"/var/run/mdsd/default_fluent.socket",
-	  WriteTimeout: 5 * time.Second,
-	  RequestAck: true}) */
+func CreateMDSDClient(dataType DataType, containerType string) {
 	mdsdfluentSocket := "/var/run/mdsd/default_fluent.socket"	
 	if containerType != "" && strings.Compare(strings.ToLower(containerType), "prometheussidecar") == 0 {
 	   mdsdfluentSocket = fmt.Sprintf("/var/run/mdsd-%s/default_fluent.socket", containerType)
 	} 
-	conn, err := net.DialTimeout("unix",
-		mdsdfluentSocket, 10*time.Second)
-	if err != nil {
-		Log("Error::mdsd::Unable to open MDSD msgp socket connection %s", err.Error())
-		//log.Fatalf("Unable to open MDSD msgp socket connection %s", err.Error())
-	} else {
-		Log("Successfully created MDSD msgp socket connection: %s", mdsdfluentSocket)
-		MdsdMsgpUnixSocketClient = conn
+	switch dataType {
+		case ContainerLogV2:
+			if MdsdMsgpUnixSocketClient != nil {
+				MdsdMsgpUnixSocketClient.Close()
+				MdsdMsgpUnixSocketClient = nil
+			}
+			/*conn, err := fluent.New(fluent.Config{FluentNetwork:"unix",
+			FluentSocketPath:"/var/run/mdsd/default_fluent.socket",
+			WriteTimeout: 5 * time.Second,
+			RequestAck: true}) */
+			conn, err := net.DialTimeout("unix",
+				mdsdfluentSocket, 10*time.Second)
+			if err != nil {
+				Log("Error::mdsd::Unable to open MDSD msgp socket connection for ContainerLogV2 %s", err.Error())
+				//log.Fatalf("Unable to open MDSD msgp socket connection %s", err.Error())
+			} else {
+				Log("Successfully created MDSD msgp socket connection for ContainerLogV2: %s", mdsdfluentSocket)
+				MdsdMsgpUnixSocketClient = conn
+			}
+		case KubeMonAgentEvents:
+			if MdsdKubeMonMsgpUnixSocketClient != nil {
+				MdsdKubeMonMsgpUnixSocketClient.Close()
+				MdsdKubeMonMsgpUnixSocketClient = nil
+			}
+			conn, err := net.DialTimeout("unix",
+				mdsdfluentSocket, 10*time.Second)
+			if err != nil {
+				Log("Error::mdsd::Unable to open MDSD msgp socket connection for KubeMon events %s",  err.Error())
+				//log.Fatalf("Unable to open MDSD msgp socket connection %s", err.Error())
+			} else {
+				Log("Successfully created MDSD msgp socket connection for KubeMon events:%s", mdsdfluentSocket)
+				MdsdKubeMonMsgpUnixSocketClient = conn
+			}
+	    case InsightsMetrics:
+			if MdsdInsightsMetricsMsgpUnixSocketClient != nil {
+				MdsdInsightsMetricsMsgpUnixSocketClient.Close()
+				MdsdInsightsMetricsMsgpUnixSocketClient = nil
+			}
+			conn, err := net.DialTimeout("unix",
+				mdsdfluentSocket, 10*time.Second)
+			if err != nil {
+				Log("Error::mdsd::Unable to open MDSD msgp socket connection for insights metrics %s", err.Error())
+				//log.Fatalf("Unable to open MDSD msgp socket connection %s", err.Error())
+			} else {
+				Log("Successfully created MDSD msgp socket connection for Insights metrics %s", mdsdfluentSocket)
+				MdsdInsightsMetricsMsgpUnixSocketClient = conn
+			}
 	}
 }