+++ source/plugins/ruby/KubernetesApiClient.rb
@@ -437,9 +437,6 @@ class KubernetesApiClient
             if (!container["resources"].nil? && !container["resources"].empty? && !container["resources"][metricCategory].nil? && !container["resources"][metricCategory][metricNameToCollect].nil?)
               metricValue = getMetricNumericValue(metricNameToCollect, container["resources"][metricCategory][metricNameToCollect])
 
-              # metricItem = {}
-              # metricItem["DataItems"] = []
-
               metricProps = {}
               metricProps["Timestamp"] = metricTime
               metricProps["Host"] = nodeName
@@ -450,7 +447,7 @@ class KubernetesApiClient
 
               metricCollection = {}
               metricCollection["CounterName"] = metricNametoReturn
-              metricCollection["Value"] = metricValue              
+              metricCollection["Value"] = metricValue
               
               metricProps["json_Collections"] = []
               metricCollections = []               
@@ -463,9 +460,7 @@ class KubernetesApiClient
               if (metricCategory == "limits" && @@NodeMetrics.has_key?(nodeMetricsHashKey))
                 metricValue = @@NodeMetrics[nodeMetricsHashKey]
                 #@Log.info("Limits not set for container #{clusterId + "/" + podUid + "/" + containerName} using node level limits: #{nodeMetricsHashKey}=#{metricValue} ")
-                # metricItem = {}
-                # metricItem["DataItems"] = []
-
+                               
                 metricProps = {}
                 metricProps["Timestamp"] = metricTime
                 metricProps["Host"] = nodeName
@@ -611,11 +606,11 @@ class KubernetesApiClient
           metricCollection = {}
           metricCollection["CounterName"] = metricNametoReturn
           metricCollection["Value"] = metricValue
-          metricCollections = []         
+          metricCollections = []
           metricCollections.push(metricCollection) 
          
-          metricItem["json_Collections"] = []                
-          metricItem["json_Collections"] = metricCollections.to_json         
+          metricItem["json_Collections"] = []
+          metricItem["json_Collections"] = metricCollections.to_json
          
           #push node level metrics to a inmem hash so that we can use it looking up at container level.
           #Currently if container level cpu & memory limits are not defined we default to node level limits