+++ source/plugins/ruby/in_cadvisor_perf.rb
@@ -54,7 +54,7 @@ module Fluent::Plugin
 
     def enumerate()
       currentTime = Time.now
-      time = currentTime.to_f
+      time = Fluent::Engine.now
       batchTime = currentTime.utc.iso8601
       @@istestvar = ENV["ISTEST"]
       begin
@@ -62,7 +62,7 @@ module Fluent::Plugin
         insightsMetricsEventStream = Fluent::MultiEventStream.new
         metricData = CAdvisorMetricsAPIClient.getMetrics(winNode: nil, metricTime: batchTime )
         metricData.each do |record|          
-          eventStream.add(Fluent::Engine.now, record) if record                  
+          eventStream.add(time, record) if record                  
         end
         
         if ExtensionUtils.isAADMSIAuthMode()
@@ -95,7 +95,7 @@ module Fluent::Plugin
           containerGPUusageInsightsMetricsDataItems.concat(CAdvisorMetricsAPIClient.getInsightsMetrics(winNode: nil, metricTime: batchTime))          
 
           containerGPUusageInsightsMetricsDataItems.each do |insightsMetricsRecord|
-            insightsMetricsEventStream.add(Fluent::Engine.now, insightsMetricsRecord) if insightsMetricsRecord
+            insightsMetricsEventStream.add(time, insightsMetricsRecord) if insightsMetricsRecord
           end
 
           router.emit_stream(@insightsmetricstag, insightsMetricsEventStream) if insightsMetricsEventStream