+++ source/plugins/ruby/in_kube_pvinventory.rb
@@ -62,7 +62,7 @@ module Fluent::Plugin
         telemetryFlush = false
         @pvTypeToCountHash = {}
         currentTime = Time.now
-        batchTime = currentTime.utc.iso8601
+        batchTime = currentTime.utc.iso8601           
 
         if ExtensionUtils.isAADMSIAuthMode()
           $log.info("in_kube_pvinventory::enumerate: AAD AUTH MSI MODE")             
@@ -120,7 +120,8 @@ module Fluent::Plugin
     end # end enumerate
 
     def parse_and_emit_records(pvInventory, batchTime = Time.utc.iso8601)
-      currentTime = Time.now      
+      currentTime = Time.now  
+      emitTime = Fluent::Engine.now    
       eventStream = Fluent::MultiEventStream.new
       @@istestvar = ENV["ISTEST"]
       begin
@@ -162,7 +163,7 @@ module Fluent::Plugin
 
         records.each do |record|
           if !record.nil?          
-            eventStream.add(Fluent::Engine.now, record) 
+            eventStream.add(emitTime, record) 
           end
         end