+++ ReleaseNotes.md
@@ -11,6 +11,14 @@ additional questions or comments.
 
 Note : The agent version(s) below has dates (ciprod<mmddyyyy>), which indicate the agent build dates (not release dates)
 
+
+### 05/12/2021 -
+##### Version microsoft/oms:ciprod00512021 Version mcr.microsoft.com/azuremonitor/containerinsights/ciprod:ciprod05122021 (linux)
+##### No Windows changes with this release, win-ciprod04222021 still current.
+##### Code change log
+- Upgrading oneagent to version 1.8 (only for Linux)
+- Enabling oneagent for container logs for East US 2
+
 ### 04/22/2021 -
 ##### Version microsoft/oms:ciprod04222021 Version mcr.microsoft.com/azuremonitor/containerinsights/ciprod:ciprod04222021 (linux)
 ##### Version microsoft/oms:win-ciprod04222021 Version mcr.microsoft.com/azuremonitor/containerinsights/ciprod:win-ciprod04222021 (windows)
@@ -26,7 +34,6 @@ Note : The agent version(s) below has dates (ciprod<mmddyyyy>), which indicate t
 - Bug fixes
 - Doc updates
 - Minor telemetry changes
-
 
 ### 03/26/2021 -
 ##### Version microsoft/oms:ciprod03262021 Version mcr.microsoft.com/azuremonitor/containerinsights/ciprod:ciprod03262021 (linux)